move custom widgets to widgets module
