+++ escoresheet/utils/__init__.py
@@ -1,4 +1,5 @@
 from .utils import *
 from .testing_utils import *
 from .form_fields import *
+from .form_widgets import *
 from .mixins import *
